Remove `console.log()` calls from content script
Writing to the consoles log from a content script makes it possible to
detect the use of this extension and the spoofing of the User-Agent
string.

An example i can think of is the error notifier Rollbar, its JS library
detects and uploads output written to the consoles log, which can then
be viewed on their website when investigating an error.